Creating the Reset Button
This button resets the game: generating new random colors, applying them to the squares, picking one and displaying it in the header, and resetting the h1 background color to default.
Also, the div containing the button was given a style.
When you pick the right color, h1 background color matches that color and the reset button displays a 'play again?' message.