Adding functionality to the buttons
'easy' and 'hard'  buttons when clicked, generate 3 and 6 random rgb colors respectively, assigning those colors to the squares needed for that level, pick a color and display it in the header.
'easy' button also hides the 3 squares that aren't needed.
'hard' button displays back the 3 erased squares during the easy level.