Improving styles
The span is in its own line and it's twice its size.
The body was given a font and the margin was removed.
The squares were given that rounded look, and a smoothly transition to colours.
Space was added at the top-top of the squares to separate the first  row from the buttons.
The header was centred, upper-cased, its background colour, font-weight and line-height changed, margin removed, and a padding top and bottom was added.
The selected class was given a background colour and colour.
The buttons were removed from their border and background colour, upper-cased, their height, colour, font-weight, font-size, letter spacing were changed. Also a smoothly transition of colour and background colour was added while hovering.